+++ examples/basic/src/routes.ts
@@ -2,6 +2,7 @@ import { AboutPage } from "./pages/about";
 import { HomePage } from "./pages/home";
 import { PostPage } from "./pages/post";
 import { type AnyRoute, route } from "cirrojs";
+export { initCssRegistry, getCssRegistry } from "cirrojs";
 
 // サイトのルート定義（Config Base Routing）。
 export const routes: AnyRoute[] = [
@@ -9,6 +10,7 @@ export const routes: AnyRoute[] = [
     { path: "/about", component: AboutPage },
     route({
         path: ({ slug }) => `/posts/${slug}`,
+        cssPath: "/posts/index.css",
         getStaticPaths: () => [{ slug: "hello" }, { slug: "world" }],
         component: PostPage,
     }),