+++ packages/cirrojs/src/index.ts
@@ -4,3 +4,7 @@ export type { MarkdownConfig, RenderResult, ToC } from "./markdown.tsx";
 export { createMarkdown } from "./markdown.tsx";
 export type { AnyRoute, DynamicRoute, Params, ResolvedPage, StaticRoute } from "./router.ts";
 export { expandRoutes, route, urlToFilePath } from "./router.ts";
+export type { Properties } from "./properties.ts";
+export { css, genCssFn } from "./css.ts";
+export type { CssOpt } from "./css.ts";
+export { initCssRegistry, getCssRegistry } from "./registry.ts";